requirements.txt
scrapling
curl-cffi
playwright
browserforge
python-dotenv
selectolax
httpx
